feat(registration): implementar registro de proveedores y clientes con notificaciones de bienvenida
