+++ app/backend/app/Http/Requests/Api/V1/StoreUserRegisterRequest.php
@@ -6,7 +6,7 @@ namespace App\Http\Requests\Api\V1;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class ProviderRegisterRequest extends FormRequest
+class StoreUserRegisterRequest extends FormRequest
 {
     /**
      * Determina si el usuario está autorizado para hacer esta petición.
@@ -15,6 +15,19 @@ class ProviderRegisterRequest extends FormRequest
     {
         // Registro público, no requiere autorización
         return true;
+    }
+
+    public function prepareForValidation(): void
+    {
+        if ($this->is('provider/register')){
+            $this->merge([
+                'role' => 'provider', // El rol se asignará automáticamente en el controlador
+            ]);
+        } elseif ($this->is('customer/register')) {
+            $this->merge([
+                'role' => 'user', // El rol se asignará automáticamente en el controlador
+            ]);
+        }         
     }
 
     /**
